move pagination into it's own component
